Move finatra examples to app-examples module
