+++ app-examples/src/main/resources/config/plugins-finatra-service.conf
