+++ app-examples/src/main/scala/net/spals/appbuilder/app/examples/finatra/sample/web/SampleFinatraFilter.scala
@@ -1,5 +1,6 @@
-package net.spals.appbuilder.app.finatra.sample.web
+package net.spals.appbuilder.app.examples.finatra.sample.web
 
+import com.google.common.annotations.VisibleForTesting
 import com.twitter.finagle.http.{Request, Response}
 import com.twitter.finagle.{Filter, Service, SimpleFilter}
 import com.twitter.util.Future
@@ -11,7 +12,8 @@ import net.spals.appbuilder.annotations.service.AutoBindSingleton
   * @author tkral
   */
 @AutoBindSingleton
-private[finatra] class SampleFinatraFilter extends SimpleFilter[Request, Response] {
+@VisibleForTesting
+/*private[finatra]*/ class SampleFinatraFilter extends SimpleFilter[Request, Response] {
 
   override def apply(request: Request, service: Service[Request, Response]): Future[Response] = {
     service.apply(request)
