+++ app-examples/src/main/scala/net/spals/appbuilder/app/examples/finatra/tracing/TracingFinatraController.scala
@@ -1,4 +1,4 @@
-package net.spals.appbuilder.app.finatra.tracing
+package net.spals.appbuilder.app.examples.finatra.tracing
 
 import com.twitter.finagle.http.Request
 import com.twitter.finatra.http.Controller
