+++ app-finatra-test/pom.xml
@@ -52,64 +52,8 @@
             <artifactId>catch-throwable</artifactId>
         </dependency>
         <dependency>
-            <groupId>io.opentracing</groupId>
-            <artifactId>opentracing-mock</artifactId>
-        </dependency>
-        <dependency>
             <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-executor-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-filestore-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-mapstore-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-message-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder</groupId>
-            <artifactId>spals-appbuilder-model-core</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder.plugins</groupId>
-            <artifactId>spals-appbuilder-app-finatra</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder.plugins</groupId>
-            <artifactId>spals-appbuilder-filestore-s3</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder.plugins</groupId>
-            <artifactId>spals-appbuilder-mapstore-cassandra</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder.plugins</groupId>
-            <artifactId>spals-appbuilder-mapstore-dynamodb</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder.plugins</groupId>
-            <artifactId>spals-appbuilder-mapstore-mongodb</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder.plugins</groupId>
-            <artifactId>spals-appbuilder-message-kafka</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder.plugins</groupId>
-            <artifactId>spals-appbuilder-message-kinesis</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder.plugins</groupId>
-            <artifactId>spals-appbuilder-model-protobuf</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>net.spals.appbuilder.plugins</groupId>
-            <artifactId>spals-appbuilder-monitor-lightstep</artifactId>
+            <artifactId>spals-appbuilder-app-examples</artifactId>
         </dependency>
         <dependency>
             <groupId>org.hamcrest</groupId>
