--- app-finatra-test/src/test/resources/AwsCredentials.properties
@@ -1,2 +1,0 @@
-accessKey=IZFBRDUMMYVMNQAJVEQR
-secretKey=ALle9zdW47DUMMYiSALhBwMDIEkndJt6JDVx7mt
