+++ pom.xml
@@ -350,6 +350,7 @@
                 <groupId>io.opentracing</groupId>
                 <artifactId>opentracing-mock</artifactId>
                 <version>${opentracing.version}</version>
+                <scope>test</scope>
             </dependency>
             <dependency>
                 <groupId>io.opentracing</groupId>
